requirements.txt
streamlit
openai-whisper
transformers
sentencepiece
torch
sacremoses
ffmpeg-python